// src/ui/interface.ts
export interface SurveyPage {
    name: string;
    elements: any[];
};

export interface SurveyDataSettings{
    id?: string | null;
    slug: string | null;
    visibility: 'public' | 'private';
    is_active: boolean;
    type: FormTypes;
    is_retryable: boolean;
    retry_size: number;
    is_batch_question: boolean;
    batch_question_size: number;
    passing_score: number;
    is_survey_preview: boolean;
}
export interface FormCreate extends SurveyDataSettings {
    title: string;
    description: string;
    content: string; // survey data json format string 
    created_by: string;
    updated_by: string;
    created_at: Date;
    updated_at: Date;
}
export interface SurveyData {
    title: string;
    description: string;
    pages: SurveyPage[];
    questionOrder?: string;
    showProgressBar: boolean;
    progressBarLocation: string;
    showTimer: boolean;
    timeLimit: number;
    timeLimitPerPage: number;
    firstPageIsStartPage: boolean;
    startSurveyText: string;
    completedHtml: string;
    completedHtmlOnCondition: any[];
    widthMode: string;
};